+++ src/components/inputs/code-input/code-input.tsx
@@ -180,8 +180,7 @@ export class CodeInput {
 
     if (event.data && index < (this.length - 1)) {
       // apply focus on next input
-      const nextInput = this.getInputByIndex(index + 1)
-      nextInput.focus()
+      this.focusOnNextInput(index)
     }
 
     this.handleCompletedEvent()
@@ -215,7 +214,7 @@ export class CodeInput {
       Backspace: () => {
         setTimeout(() => {
           this.focusOnPreviousInput(index)
-        }, 10)
+        }, 20)
       },
       ArrowLeft: () => {
         this.focusOnPreviousInput(index)
@@ -256,7 +255,7 @@ export class CodeInput {
       if (select) {
         setTimeout(() => {
           input.select()
-        }, 10)
+        }, 30)
       }
     }
     return input